+++ packages/app/src/design-system/theme.ts
@@ -4,5 +4,6 @@ import "@gouvfr/dsfr/dist/utility/icons/icons-user/icons-user.min.css";
 import "@gouvfr/dsfr/dist/utility/icons/icons-business/icons-business.min.css";
 import "@gouvfr/dsfr/dist/utility/icons/icons-map/icons-map.min.css";
 import "@gouvfr/dsfr/dist/utility/icons/icons-design/icons-design.min.css";
+import "@gouvfr/dsfr/dist/utility/icons/icons-editor/icons-editor.min.css";
 
 export {};